api/requirements.txt
fastapi==0.100.0
motor==3.2.0
psutil==5.9.5
python-dotenv==1.0.0
uvicorn==0.22.0
